refactor(plugin-command): add plugin-command package
