+++ .github/workflows/test packages.yml
@@ -122,3 +122,19 @@ jobs:
 
       - name: test
         run: cd packages/editor-core && npm test
+
+  test-plugin-command
+    runs-on: ubuntu-latest
+    steps:
+      - name: checkout
+        uses: actions/checkout@v2
+
+      - uses: actions/setup-node@v2
+        with:
+          node-version: '14'
+
+      - name: install
+        run: npm i && npm run setup:skip-build
+
+      - name: test
+        run: cd packages/plugin-command && npm test
