+++ packages/plugin-command/README.md
@@ -1,0 +1,11 @@
+# `@alilc/plugin-command`
+
+> TODO: description
+
+## Usage
+
+```
+const pluginCommand = require('@alilc/plugin-command');
+
+// TODO: DEMONSTRATE API
+```
