+++ packages/plugin-command/build.json
@@ -1,0 +1,9 @@
+{
+  "plugins": [
+    "@alilc/build-plugin-lce",
+    "build-plugin-fusion",
+    ["build-plugin-moment-locales", {
+      "locales": ["zh-cn"]
+    }]
+  ]
+}
