+++ packages/plugin-command/build.test.json
@@ -1,0 +1,19 @@
+{
+  "plugins": [
+    [
+      "@alilc/build-plugin-lce",
+      {
+        "filename": "editor-preset-vision",
+        "library": "LowcodeEditor",
+        "libraryTarget": "umd",
+        "externals": {
+          "react": "var window.React",
+          "react-dom": "var window.ReactDOM",
+          "prop-types": "var window.PropTypes",
+          "rax": "var window.Rax"
+        }
+      }
+    ],
+    "@alilc/lowcode-test-mate/plugin/index.ts"
+  ]
+}
